Remove underscore for content type modules
