+++ src/senaite/core/browser/content/configure.zcml
@@ -6,9 +6,9 @@
 
   <!-- Result Interpretation Templates View -->
   <browser:page
-    for="senaite.core.content.interpretation_template.IInterpretationTemplates"
+    for="senaite.core.content.interpretationtemplates.IInterpretationTemplates"
     name="view"
-    class=".interpretation_templates.InterpretationTemplatesView"
+    class=".interpretationtemplates.InterpretationTemplatesView"
     permission="senaite.core.permissions.ManageBika"
     layer="senaite.core.interfaces.ISenaiteCore"
   />
