+++ src/senaite/core/browser/content/interpretationtemplates.py
