+++ src/senaite/core/content/interpretationtemplates.py
@@ -1,0 +1,42 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of SENAITE.CORE.
+#
+# SENAITE.CORE is free software: you can redistribute it and/or modify it under
+# the terms of the GNU General Public License as published by the Free Software
+# Foundation, version 2.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+# details.
+#
+# You should have received a copy of the GNU General Public License along with
+# this program; if not, write to the Free Software Foundation, Inc., 51
+# Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+#
+# Copyright 2018-2020 by it's authors.
+# Some rights reserved, see README and LICENSE.
+
+from plone.dexterity.content import Container
+from plone.supermodel import model
+
+from senaite.core import PloneMessageFactory as _PMF
+
+from zope import schema
+from zope.interface import implementer
+
+
+class IInterpretationTemplates(model.Schema):
+    """Results Interpretation Templates folder interface
+    """
+    title = schema.TextLine(
+        title=_PMF(u'label_title', default=u'Title'),
+        required=True)
+
+
+@implementer(IInterpretationTemplates)
+class InterpretationTemplates(Container):
+    """Results Interpretation Templates folder
+    """
+    pass
